reorg, sender init

+++ bot/src/redis.ts
@@ -1,11 +1,29 @@
 import { DiscordGuild } from "discordeno";
+import { connect, Redis } from "redis";
 import { deserialize, serialize } from "~json/initiator.ts";
-import { logger } from "../utils/logger.ts";
-import { redis } from "./mod.ts";
+import { config } from "./config.ts";
+import { logger } from "./utils/logger.ts";
 
-export const redisGet = async <TData>(
-  key: string
-): Promise<TData | undefined> => {
+export let redis: Redis | null = null;
+
+export const connectRedis = async () => {
+  if (redis) {
+    return;
+  }
+
+  redis = await connect({
+    hostname: config.REDISHOST,
+    port: config.REDISPORT,
+    password: config.REDISPASSWORD,
+    username: config.REDISUSER,
+  });
+
+  logger.info("Connected to Redis");
+};
+
+export const redisGet = async <TData>(key: string): Promise<TData | undefined> => {
+  if (!redis) throw new Error("Not connected to Redis");
+
   try {
     const serializedData = await redis.get(key);
 
@@ -26,6 +44,8 @@ export const redisSet = async (
   // deno-lint-ignore no-explicit-any
   data: any
 ): Promise<boolean> => {
+  if (!redis) throw new Error("Not connected to Redis");
+
   try {
     const serializedData = await serialize(data);
 
@@ -40,6 +60,8 @@ export const redisSet = async (
 };
 
 export const redisDelete = async (...keys: string[]): Promise<boolean> => {
+  if (!redis) throw new Error("Not connected to Redis");
+
   try {
     await redis.unlink(...keys);
 
@@ -51,15 +73,12 @@ export const redisDelete = async (...keys: string[]): Promise<boolean> => {
   }
 };
 
-export const redisGuildId = (guildId: bigint | string): string =>
-  `guild:${guildId}`;
+export const redisGuildId = (guildId: bigint | string): string => `guild:${guildId}`;
 export const redisChannelId = (guildId: bigint, channelId: bigint): string =>
   `channel:${guildId}/${channelId}`;
 
 export const setGuildToCache = (guild: DiscordGuild): Promise<boolean> =>
   redisSet(redisGuildId(guild.id), guild);
 
-export const getCachedGuild = (
-  guildId: bigint
-): Promise<DiscordGuild | undefined> =>
+export const getCachedGuild = (guildId: bigint): Promise<DiscordGuild | undefined> =>
   redisGet<DiscordGuild>(redisGuildId(guildId));
